+++ CHANGELOG.md
@@ -295,6 +295,7 @@ Yii Framework 2 Change Log
 - New: Added `yii\web\GroupUrlRule` (qiangxue)
 - New: Added `yii\filters\RateLimiter` (qiangxue)
 - New: Added various authentication methods, including `HttpBasicAuth`, `HttpBearerAuth`, `QueryParamAuth`, and `CompositeAuth` (qiangxue)
+- New: Added `HtmlResponseFormatter` and `JsonResponseFormatter` (qiangxue)
 
 2.0.0-alpha, December 1, 2013
 -----------------------------